network lib and use sp1_sdk ProofFromNetwork
