+++ hierophant/network-lib/src/lib.rs
@@ -1,17 +1,12 @@
+pub mod messages;
+pub mod protocol;
+
 use alloy_primitives::B256;
 use serde::{Deserialize, Serialize};
+use sp1_sdk::ProofFromNetwork;
 use sp1_sdk::network::proto::network::ExecutionStatus;
-use sp1_sdk::{
-    SP1Proof, SP1ProofWithPublicValues, SP1PublicValues, SP1Stdin,
-    network::proto::network::ProofMode,
-};
+use sp1_sdk::{SP1ProofWithPublicValues, SP1Stdin, network::proto::network::ProofMode};
 use std::{cmp::Ordering, fmt::Display};
-
-pub const REGISTER_CONTEMPLANT_ENDPOINT: &str = "register_contemplant";
-// Increment this whenever there is a breaking change in the contemplant
-// This is to ensure the contemplant is on the same version as the Hierophant it's
-// connecting to
-pub const CONTEMPLANT_VERSION: &str = "6.0.0";
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct WorkerRegisterInfo {
@@ -33,44 +28,6 @@ impl Display for WorkerRegisterInfo {
             "{} CONTEMPLANT_VERSION {}{}",
             self.name, self.contemplant_version, magister_info
         )
-    }
-}
-
-#[derive(Serialize, Deserialize)]
-pub enum FromContemplantMessage {
-    // sent from contemplant to hierophant on startup
-    Register(WorkerRegisterInfo),
-    // sends proof_status responses to the hierophant
-    ProofStatusResponse(B256, Option<ContemplantProofStatus>),
-    Heartbeat,
-}
-
-impl Display for FromContemplantMessage {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let msg = match self {
-            Self::Register(_) => "Register",
-            Self::ProofStatusResponse(_, _) => "ProofStatusResponse",
-            Self::Heartbeat => "Heartbeat",
-        };
-        write!(f, "{msg}")
-    }
-}
-
-#[derive(Serialize, Deserialize)]
-pub enum FromHierophantMessage {
-    // sent from hierophant to contemplant to start working on a new proof
-    ProofRequest(ContemplantProofRequest),
-    // sent from hierophant to contemplant to get the status of a proof
-    ProofStatusRequest(B256),
-}
-
-impl Display for FromHierophantMessage {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let msg = match self {
-            Self::ProofRequest(_) => "ProofRequest",
-            Self::ProofStatusRequest(_) => "ProofStatusRequest",
-        };
-        write!(f, "{msg}")
     }
 }
 
@@ -123,8 +80,10 @@ impl ContemplantProofStatus {
         };
         self.progress = updated_progress;
     }
+}
 
-    pub fn default() -> Self {
+impl Default for ContemplantProofStatus {
+    fn default() -> Self {
         Self {
             execution_status: ExecutionStatus::Unexecutable.into(),
             proof: None,
@@ -165,7 +124,7 @@ impl Display for ContemplantProofStatus {
 pub enum ProgressUpdate {
     Execution(u64),     // 0 to 100
     Serialization(u64), // 0 to 100
-    Done,               // Finished with a count of serialization shards.
+    Done,               // Finished
 }
 
 impl Display for ProgressUpdate {
@@ -198,6 +157,7 @@ impl PartialOrd for ProgressUpdate {
     }
 }
 
+// Progress order goes Execution(0) -> Execution(100) -> Serialization(0) -> Serialization(100) -> Done
 impl Ord for ProgressUpdate {
     fn cmp(&self, other: &Self) -> Ordering {
         match (self, other) {
@@ -217,51 +177,13 @@ impl Ord for ProgressUpdate {
     }
 }
 
-/// A proof generated by the SP1 RISC-V zkVM bundled together with the public values and the
-/// version.
-/*
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SP1ProofWithPublicValues {
-    /// The raw proof generated by the SP1 RISC-V zkVM.
-    pub proof: SP1Proof,
-    /// The public values generated by the SP1 RISC-V zkVM.
-    pub public_values: SP1PublicValues,
-    /// The version of the SP1 RISC-V zkVM (not necessary but useful for detecting version
-    /// mismatches).
-    pub sp1_version: String,
-    /// The integrity proof generated by the TEE server.
-    pub tee_proof: Option<Vec<u8>>,
-}
-*/
-// The proof generated by the prover network.
-//
-// Since [`bincode`] is not self describing, it cannot handle "nullable" optional values.
-//
-// This is the same exact type as ProofFromNetwork in sp1_sdk but their's isn't public
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ProofFromNetwork {
-    proof: SP1Proof,
-    public_values: SP1PublicValues,
-    sp1_version: String,
-}
-
-impl From<ProofFromNetwork> for SP1ProofWithPublicValues {
-    fn from(value: ProofFromNetwork) -> Self {
-        Self {
-            proof: value.proof,
-            public_values: value.public_values,
-            sp1_version: value.sp1_version,
-            tee_proof: None,
-        }
-    }
-}
-
-impl From<SP1ProofWithPublicValues> for ProofFromNetwork {
-    fn from(value: SP1ProofWithPublicValues) -> Self {
-        Self {
-            proof: value.proof,
-            public_values: value.public_values,
-            sp1_version: value.sp1_version,
-        }
+// helper function
+// sp1_sdk doesn't impl From<SP1ProofWithPublicValues> for ProofFromNetwork so we have to make a
+// fake impl.
+pub fn to_proof_from_network(p: SP1ProofWithPublicValues) -> ProofFromNetwork {
+    ProofFromNetwork {
+        proof: p.proof,
+        public_values: p.public_values,
+        sp1_version: p.sp1_version,
     }
 }
