+++ hierophant/network-lib/src/protocol.rs
@@ -1,0 +1,6 @@
+pub const REGISTER_CONTEMPLANT_ENDPOINT: &str = "register_contemplant";
+
+// Increment this whenever there is a breaking change in the contemplant
+// This is to ensure the contemplant is on the same version as the Hierophant it's
+// connecting to
+pub const CONTEMPLANT_VERSION: &str = "6.0.0";
